feat : 회원 로그인 추가

+++ src/main/java/com/user/repository/MemberDetailRepository.java
@@ -5,6 +5,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface MemberDetailRepository extends JpaRepository<MemberDetail, Integer> {
+public interface MemberDetailRepository extends JpaRepository<MemberDetail, Long> {
+
 
 }
